Made checklists save state and appear on home screen
