+++ to_dont_list/lib/main.dart
@@ -1,6 +1,7 @@
 // Started with https://docs.flutter.dev/development/ui/widgets-intro
 import 'package:flutter/material.dart';
 import 'package:to_dont_list/objects/checklist_item.dart';
+import 'package:to_dont_list/settings.dart';
 import 'package:to_dont_list/widgets/checklist_list_item.dart';
 import 'package:to_dont_list/widgets/checklist_dialog.dart';
 import 'checklist_data.dart';
@@ -64,11 +65,17 @@ class _ToDoListState extends State<ToDoList> {
             setState(() {
               checklists.add(newChecklistName);
               checklistSteps[newChecklistName] = [];
+              selected.add(newChecklistName);
             });
           },
         );
       },
     );
+  }
+
+  void _setChecklist(){
+    setState(() {
+    });
   }
 
   @override
@@ -79,11 +86,11 @@ class _ToDoListState extends State<ToDoList> {
         actions: [IconButton(onPressed: () async {
             await Navigator.of(context).push(
               MaterialPageRoute(
-                builder: (context) => SettingsScreen()));} , icon: Icon(Icons.settings))],
+                builder: (context) => SettingsScreen(passedlist: [displaychecklists], setChecklist: _setChecklist,)));} , icon: Icon(Icons.settings))],
       ),
       body: ListView(
         padding: const EdgeInsets.symmetric(vertical: 8.0),
-        children: displaychecklists.map((checklistName) {
+        children: selected.map((checklistName) {
           return ListTile(
             leading: const CircleAvatar(
               backgroundColor: Colors.red,
@@ -109,46 +116,8 @@ class _ToDoListState extends State<ToDoList> {
   }
 }
 
-class SettingsScreen extends StatefulWidget {
-  const SettingsScreen({super.key});
-  
 
-  @override
-  State<SettingsScreen> createState() => _SettingsScreenState();
-}
 
-class _SettingsScreenState extends State<SettingsScreen> {
-  bool? value = false;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Settings'),
-        ),
-    
-      body: ListView(
-        padding: const EdgeInsets.symmetric(vertical: 8.0),
-        children: checklists.map((checklistName) {
-          return ListTile(
-            leading: null,
-            title: Text(checklistName),
-            onTap: () {},
-            trailing: 
-                      Checkbox(
-                        value: value,
-                        onChanged: (bool? newValue) {
-                          setState(() {
-                            value = newValue;
-                            
-                          });
-                        },
-                      ),
-          );
-        }).toList(),
-      ),
-    );
-  }
-  }
 
 
 class ChecklistDetailsScreen extends StatefulWidget {
